app/fastapi_kafka/requirements.txt
# aiokafka==0.12.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
asyncpg==0.31.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
fastapi==0.124.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
kafka==1.3.5
packaging==25.0
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
uvloop==0.22.1
aws-msk-iam-sasl-signer-python==1.0.2
kafka-python==2.3.0
kafka-python-ng==2.2.3